% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerExploreData-package.R
\docType{package}
\name{cerExploreData-package}
\alias{cerExploreData}
\alias{cerExploreData-package}
\title{cerExploreData: Summary Statistics and Visualization}
\description{
Summary Statistics and Visualization.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
