% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSummary.R
\name{renderSummary}
\alias{renderSummary}
\title{Render Summary}
\usage{
renderSummary(
  outDir = getwd(),
  outFileName = "Summary",
  outFormat = "HTML",
  envir = new.env(),
  quiet = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{outDir}{\code{character(1)}. The directory where the output document
and intermediary files are created.}

\item{outFileName}{\code{character(1)}. File name (without extension) of the
output document to be created.}

\item{outFormat}{\code{character(1)}, not case-sensitive. \code{"html"},
\code{"pdf"}, or \code{"rmd"}.}

\item{envir}{\code{environment}. The environment in which the code chunks are
to be evaluated. See the argument of the same name in
\code{\link[rmarkdown:render]{rmarkdown::render()}}. Set this to
\code{globalenv()} and \code{sections} to \code{NULL} to load an preprocess
data in your global environment during development.}

\item{quiet}{\code{logical(1)}. Suppress printing during rendering?}

\item{debug}{\code{logical(1)}. Should the debug Rmd be rendered instead of
the main one?}

\item{...}{YAML parameters, see below.}
}
\value{
The value returned by
\code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Renders the the Rmd summary for a variable in a csv that has a region and a
time column.
}
\section{YAML Parameters}{

\describe{
\item{\code{dataFilePath}}{
\code{character(1)}.
Path to a csv file.}
\item{\code{variableName}}{
\code{character(1)}.
The name of the column in the csv file \code{dataFilePath} to be shown.}
\item{\code{regionName}}{
\code{character(1)}.
Default: \code{"GID_1"}.
The name of the region column in the csv file \code{dataFilePath}.}
\item{\code{regionNameDelim}}{
\code{character(1)}.
Default: \code{"."}.
Delimiter to split the \code{regionName} into country and subRegionId.}
\item{\code{regionNamePartNames}}{
\code{character(2)}.
Default: \code{c("country", "subRegionId")}.
Names of the parts of the \code{regionName} after split by \code{regionNameDelim}.}
\item{\code{countryName}}{
\code{character(1)}.
Default: \code{"country"}.
Name of the country column in the csv file \code{dataFilePath}.}
\item{\code{timeName}}{
\code{character(1)}.
Default: \code{"year"}.
The name of the time column in the csv file \code{dataFilePath}.}
\item{\code{timeRange}}{
\code{numeric(2)} or \code{NULL}.
Default: \code{c(1950, 2050)}.
The maximum time range to be shown.}
\item{\code{shapeRegionFilePath}}{
\code{character(1)}.
Path to a shapefile with region geometries.}
\item{\code{shapeRegionLayer}}{
\code{NULL} or \code{character(1)}.
Default: \code{NULL}.
The name of the layer in the shapefile \code{shapeRegionFilePath} with region geometries.}
\item{\code{shapeRegionName}}{
\code{character(1)}.
Default: \code{"GID_1"}.
The name of the region column in the shapefile \code{shapeRegionFilePath}.}
\item{\code{shapeCountryFilePath}}{
\code{character(1)}.
Path to a shapefile with country geometries.}
\item{\code{shapeCountryLayer}}{
\code{NULL} or \code{character(1)}.
Default: \code{NULL}.
The name of the layer in the shapefile \code{shapeCountryFilePath} with country geometries.}
\item{\code{shapeCountryName}}{
\code{character(1)}.
Default: \code{"GID_1"}.
The name of the country column in the shapefile \code{shapeCountryFilePath}.}
\item{\code{regionRegex}}{
\code{character(1)} or \code{NULL}.
Default: \code{NULL}.
A regular expression to filter the regions.}
\item{\code{countryRegex}}{
\code{character(1)} or \code{NULL}.
Default: \code{NULL}.
A regular expression to filter the countries.}
\item{\code{aggregateFunctions}}{
\code{list(...)}.
Default: \code{list(mean = \(x) mean(x, na.rm=TRUE), sd = \(x) sd(x, na.rm=TRUE))}.
A named list of functions to aggregate values from a group. The first one (typically mean or sum) also marks the no-aggregation statistic.}
\item{\code{transformations}}{
\code{list(...)}.
Default: \code{list(mean = c("identity", "log10"), sd = "identity")}.
A named list of character vectors with the same names as \code{aggregateFunctions}. The transformations applied to the scales of the (aggregated) variable for plotting.}
\item{\code{createMaps}}{
\code{logical(1)}.
Default: \code{TRUE}.
Should maps be created? This might take some minutes.}
\item{\code{figWidth, figHeight}}{
\code{numeric(1)}.
Default: \code{10} and \code{6}, respectively.
Size of plots in inches.}
\item{\code{warning}}{
\code{logical(1)}.
Default: \code{FALSE}.
Show warnings in output?}
}
}

\author{
Christof Schoetz
}
